requirements.txt
flask>=2.0.1
python-binance>=1.0.16
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=0.19.0 